+++ design-system/button/examples/95-split-button-nested-modal.tsx
@@ -1,7 +1,13 @@
 import React, { useState } from 'react';
 
 import Button, { SplitButton } from '@atlaskit/button/new';
-import Modal, { ModalBody, ModalFooter, ModalTransition } from '@atlaskit/modal-dialog';
+import Modal, {
+	ModalBody,
+	ModalFooter,
+	ModalHeader,
+	ModalTitle,
+	ModalTransition,
+} from '@atlaskit/modal-dialog';
 
 export default () => {
 	const [isOpen, setIsOpen] = useState(true);
@@ -13,6 +19,9 @@ export default () => {
 				<ModalTransition>
 					{isOpen && (
 						<Modal onClose={() => setIsOpen(true)}>
+							<ModalHeader hasCloseButton>
+								<ModalTitle>Call to action</ModalTitle>
+							</ModalHeader>
 							<ModalBody>Call to action buttons below should be neutral grey (not blue)</ModalBody>
 							<ModalFooter>
 								<Button onClick={() => setIsOpen(false)}>Cancel</Button>