+++ linking-platform/linking-types/src/supported-feature.ts
@@ -1,4 +1,4 @@
 /**
  * Describes the supported feature variants
  */
-export type SupportedFeature = 'AISummary' | 'AutomationAction' | 'RelatedLinks';
+export type SupportedFeature = 'AISummary' | 'AutomationAction' | 'RelatedLinks' | 'ExportBlocked';