+++ design-system/button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/button
 
+## 22.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 22.0.2
 
 ### Patch Changes